requirements.txt
motor
python-dotenv
pymongo
discord.py
bson